Renamed resultstore test to resultsservice test
